+++ src/main/java/be/robbevanherck/javafraggenescan/repositories/NonCodingEmissionRepository.java
@@ -3,6 +3,7 @@ package be.robbevanherck.javafraggenescan.repositories;
 import be.robbevanherck.javafraggenescan.entities.AminoAcid;
 import be.robbevanherck.javafraggenescan.entities.Pair;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
@@ -10,7 +11,7 @@ import java.util.Scanner;
 /**
  * Repository for the noncoding file
  */
-public class NonCodingEmissionRepository extends CGDependentRepository<Map<Pair<AminoAcid>, Double>> {
+public class NonCodingEmissionRepository extends CGDependentRepository<Map<Pair<AminoAcid>, BigDecimal>> {
     /**
      * Create a new NonCodingEmissionRepository
      */
@@ -21,10 +22,10 @@ public class NonCodingEmissionRepository extends CGDependentRepository<Map<Pair<
     private static NonCodingEmissionRepository instance;
 
     @Override
-    protected Map<Pair<AminoAcid>, Double> readOneBlock(Scanner s) {
-        Map<Pair<AminoAcid>, Double> ret = new HashMap<>();
+    protected Map<Pair<AminoAcid>, BigDecimal> readOneBlock(Scanner s) {
+        Map<Pair<AminoAcid>, BigDecimal> ret = new HashMap<>();
         for (int aminoAcidInteger = 0; aminoAcidInteger < 16; aminoAcidInteger++) {
-                ret.put(createDinucleotide(aminoAcidInteger), s.nextDouble());
+                ret.put(createDinucleotide(aminoAcidInteger), BigDecimal.valueOf(s.nextDouble()));
         }
         return ret;
     }
