+++ src/test/java/be/robbevanherck/javafraggenescan/dummies/DummyAcidsViterbiStep.java
@@ -4,6 +4,7 @@ import be.robbevanherck.javafraggenescan.entities.AminoAcid;
 import be.robbevanherck.javafraggenescan.entities.HMMState;
 import be.robbevanherck.javafraggenescan.entities.ViterbiStep;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -47,7 +48,7 @@ public class DummyAcidsViterbiStep extends ViterbiStep {
     }
 
     @Override
-    public double getProbabilityFor(HMMState state) {
-        return 0.7;
+    public BigDecimal getProbabilityFor(HMMState state) {
+        return BigDecimal.valueOf(0.7);
     }
 }
