+++ src/test/java/be/robbevanherck/javafraggenescan/dummies/DummyMatchTransition.java
@@ -3,11 +3,13 @@ package be.robbevanherck.javafraggenescan.dummies;
 import be.robbevanherck.javafraggenescan.entities.*;
 import be.robbevanherck.javafraggenescan.transitions.MatchTransition;
 
+import java.math.BigDecimal;
+
 public class DummyMatchTransition extends MatchTransition {
-    public PathProbability lowerPathProbability = new PathProbability(HMMState.INSERT_1, 0.1);
-    public PathProbability lowPathProbability = new PathProbability(HMMState.MATCH_1, 0.2);
-    public PathProbability middlePathProbability = new PathProbability(HMMState.MATCH_2, 0.3);
-    public PathProbability highPathProbability = new PathProbability(HMMState.START, 0.4);
+    public PathProbability lowerPathProbability = new PathProbability(HMMState.INSERT_1, BigDecimal.valueOf(0.1));
+    public PathProbability lowPathProbability = new PathProbability(HMMState.MATCH_1, BigDecimal.valueOf(0.2));
+    public PathProbability middlePathProbability = new PathProbability(HMMState.MATCH_2, BigDecimal.valueOf(0.3));
+    public PathProbability highPathProbability = new PathProbability(HMMState.START, BigDecimal.valueOf(0.4));
 
     /**
      * Create a new Dummy MatchTransition (all calculation methods are mocked)
